Конфликт разрешён

+++ src/SalesManager.java
@@ -14,9 +14,9 @@ public class SalesManager {
         }
         return max;
     }
-    public int min() {
-        int min = 100;
-        for (int sale : sales) {
+    public long min() {
+        long min = 100;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -24,8 +24,8 @@ public class SalesManager {
         return min;
     }
 
-    public int average() {
-        int sum = 0;
+    public long average() {
+        long sum = 0;
         if (sales.length == 7) {
             for (long sale : sales) {
                 sum += sale;
